Implement Form Scopes
Forms are available to users depending on the form's scope.
A form with a scope of `group` can be seen by the users
associated with the groups that the form is associated with. This is
rule is similar to teams and programs. A form with a `universal` scope
is available to all users while a form with `self` scope is available only to
the user that created the form.